test_ir_consistency created
In summary, I ran multiple tests on the features of all the spectra and cut the files that exceeded a maximum length when controlled for noise. I also added a visualizer.

refinement_200-1500-100.txt is a matrix of three rows, and each column contains a valid data point's directory with the structure

-PEMDAS_B351Proj
---new_spectra_all
-----Nitrogenic
-----NoNitrogen
-test_ir_consistency

this should allow us to start training